Commit Alpha 1.0.0

+++ app/src/main/java/com/worldwidewealth/termtem/model/LoginResponseModel.java
@@ -10,14 +10,14 @@ public class LoginResponseModel {
     private boolean CASHSERVICE;
     private String AGENTID;
     private String USERID;
-    private Double BALANCE;
+    private float BALANCE;
     private List<UserMenuModel> usermenu;
 
-    public Double getBALANCE() {
+    public float getBALANCE() {
         return BALANCE;
     }
 
-    public void setBALANCE(Double BALANCE) {
+    public void setBALANCE(float BALANCE) {
         this.BALANCE = BALANCE;
     }
